Add files via upload
In this statistical project, we received a data set consisting of 18 variables and 7933 observations, with the goal of trying to predict and interpret total wealth. Using various financial, retirement, home ownership, and age variables, and also binary covariates, I tried to develop the optimal prediction model that would best predict the outcome of total wealth using the provided data set and various models and testing procedures learned throughout the course.